+++ supportedBackends/google/pipelines/common/src/main/scala/cromwell/backend/google/pipelines/common/ProjectLabels.scala
@@ -1,3 +1,3 @@
 package cromwell.backend.google.pipelines.common
 
-case class ProjectLabels(labels: Map[String, String])
+final case class ProjectLabels(labels: Map[String, String])