+++ supportedBackends/google/pipelines/v2alpha1/src/main/scala/cromwell/backend/google/pipelines/v2alpha1/api/MonitoringAction.scala
@@ -10,12 +10,12 @@ trait MonitoringAction {
                             )(implicit gcsTransferConfiguration: GcsTransferConfiguration): List[Action] = {
 
     val monitoringImageScriptActions =
-      createPipelineParameters.monitoringImageScript match {
+      createPipelineParameters.monitoringImage.monitoringImageScriptOption match {
         case Some(script) =>
           val localizeScriptAction =
             ActionBuilder.monitoringImageScriptAction(
               script,
-              createPipelineParameters.monitoringImageScriptContainerPath,
+              createPipelineParameters.monitoringImage.monitoringImageScriptContainerPath,
               mounts,
             )
           val describeLocalizeScriptAction =
@@ -28,30 +28,32 @@ trait MonitoringAction {
       }
 
     val monitoringImageActions =
-      createPipelineParameters.monitoringImage match {
-      case Some(image) =>
+      createPipelineParameters.monitoringImage.monitoringImageOption match {
+        case Some(image) =>
 
-        val monitoringImage = image
-        val monitoringImageCommand = createPipelineParameters.monitoringImageCommand
-        val monitoringImageEnvironment = createPipelineParameters.monitoringImageEnvironment
+          val monitoringImage = image
+          val monitoringImageCommand = createPipelineParameters.monitoringImage.monitoringImageCommand
+          val monitoringImageEnvironment = createPipelineParameters.monitoringImage.monitoringImageEnvironment
 
-        val monitoringAction = ActionBuilder.monitoringAction(
-          monitoringImage,
-          monitoringImageCommand,
-          monitoringImageEnvironment(mounts.map(_.getPath)),
-          mounts,
-        )
-        val describeMonitoringAction = ActionBuilder.describeDocker("monitoring action", monitoringAction)
-        List(describeMonitoringAction, monitoringAction)
+          val monitoringAction = ActionBuilder.monitoringAction(
+            monitoringImage,
+            monitoringImageCommand,
+            monitoringImageEnvironment(mounts.map(_.getPath)),
+            mounts,
+          )
 
-      case None => Nil
-    }
+          val describeMonitoringAction = ActionBuilder.describeDocker("monitoring action", monitoringAction)
+
+          List(describeMonitoringAction, monitoringAction)
+
+        case None => Nil
+      }
 
     monitoringImageScriptActions ++ monitoringImageActions
   }
 
   def monitoringShutdownActions(createPipelineParameters: CreatePipelineParameters): List[Action] = {
-    createPipelineParameters.monitoringImage match {
+    createPipelineParameters.monitoringImage.monitoringImageOption match {
       case Some(_) =>
         val terminationAction = ActionBuilder.monitoringTerminationAction()